implementing the SMPP5 RA
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@14086 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ resources/smpp5/events/src/main/java/net/java/slee/resources/smpp/pdu/SmppRequest.java
@@ -1,5 +1,6 @@
 package net.java.slee.resources.smpp.pdu;
 
+
 /**
  * 
  * @author amit bhayani
@@ -7,39 +8,40 @@ package net.java.slee.resources.smpp.pdu;
  */
 public interface SmppRequest extends PDU {
 
-	public static final long BIND_RECEIVER = 0x00000001;
-	public static final long BIND_TRANSMITTER = 0x00000002;
-	public static final long QUERY_SM = 0x00000003;
-	public static final long SUBMIT_SM = 0x00000004;
-	public static final long DELIVER_SM = 0x00000005;
-	public static final long UNBIND = 0x00000006;
-	public static final long REPLACE_SM = 0x00000007;
-	public static final long CANCEL_SM = 0x00000008;
-	public static final long BIND_TRANSCEIVER = 0x00000009;
-	public static final long OUTBIND = 0x0000000B;
-	public static final long ENQUIRE_LINK = 0x00000015;
-	public static final long SUBMIT_MULTI = 0x00000021;
-	public static final long ALERT_NOTIFICATION = 0x00000102;
-	public static final long DATA_SM = 0x00000103;
-	public static final long BROADCAST_SM = 0x00000111;
-	public static final long QUERY_BROADCAST_SM = 0x00000112;
-	public static final long CANCEL_BROADCAST_SM = 0x00000113;
-	public static final long GENERIC_NACK = 0x80000000;
-	public static final long BIND_RECEIVER_RESP = 0x80000001;
-	public static final long BIND_TRANSMITTER_RESP = 0x80000002;
-	public static final long QUERY_SM_RESP = 0x80000003;
-	public static final long SUBMIT_SM_RESP = 0x80000004;
-	public static final long DELIVER_SM_RESP = 0x80000005;
-	public static final long UNBIND_RESP = 0x80000006;
-	public static final long REPLACE_SM_RESP = 0x80000007;
-	public static final long CANCEL_SM_RESP = 0x80000008;
-	public static final long BIND_TRANSCEIVER_RESP = 0x80000009;
-	public static final long ENQUIRY_LINK_RESP = 0x80000015;
-	public static final long SUBMIT_MULTI_RESP = 0x80000021;
-	public static final long DATA_SM_RESP = 0x80000103;
-	public static final long BORADCAST_SM_RESP = 0x80000111;
-	public static final long QUERY_BROADCAST_SM_RESP = 0x80000112;
-	public static final long CANCEL_BROADCAST_SM_RESP = 0x80000113;
+	public static final int BIND_RECEIVER = 0x00000001;
+	public static final int BIND_TRANSMITTER = 0x00000002;
+	public static final int QUERY_SM = 0x00000003;
+	public static final int SUBMIT_SM = 0x00000004;
+	public static final int DELIVER_SM = 0x00000005;
+	public static final int UNBIND = 0x00000006;
+	public static final int REPLACE_SM = 0x00000007;
+	public static final int CANCEL_SM = 0x00000008;
+	public static final int BIND_TRANSCEIVER = 0x00000009;
+	public static final int OUTBIND = 0x0000000B;
+	public static final int ENQUIRE_LINK = 0x00000015;
+	public static final int SUBMIT_MULTI = 0x00000021;
+	public static final int ALERT_NOTIFICATION = 0x00000102;
+	public static final int DATA_SM = 0x00000103;
+	public static final int BROADCAST_SM = 0x00000111;
+	public static final int QUERY_BROADCAST_SM = 0x00000112;
+	public static final int CANCEL_BROADCAST_SM = 0x00000113;
+	public static final int GENERIC_NACK = 0x80000000;
+	public static final int BIND_RECEIVER_RESP = 0x80000001;
+	public static final int BIND_TRANSMITTER_RESP = 0x80000002;
+	public static final int QUERY_SM_RESP = 0x80000003;
+	public static final int SUBMIT_SM_RESP = 0x80000004;
+	public static final int DELIVER_SM_RESP = 0x80000005;
+	public static final int UNBIND_RESP = 0x80000006;
+	public static final int REPLACE_SM_RESP = 0x80000007;
+	public static final int CANCEL_SM_RESP = 0x80000008;
+	public static final int BIND_TRANSCEIVER_RESP = 0x80000009;
+	public static final int ENQUIRY_LINK_RESP = 0x80000015;
+	public static final int SUBMIT_MULTI_RESP = 0x80000021;
+	public static final int DATA_SM_RESP = 0x80000103;
+	public static final int BORADCAST_SM_RESP = 0x80000111;
+	public static final int QUERY_BROADCAST_SM_RESP = 0x80000112;
+	public static final int CANCEL_BROADCAST_SM_RESP = 0x80000113;
 
 	SmppResponse createSmppResponseEvent(int status);
+	
 }
